+++ client/src/components/AddItem.js
@@ -8,7 +8,7 @@ import {
   Button,
 } from "@material-ui/core";
 import { makeStyles } from "@material-ui/core/styles";
-import AddItemPopup from "pages/AddItemPopup";
+import AddItemPopup from "components/AddItemPopup";
 
 const useStyles = makeStyles((theme) => ({
   dashboardAddItem: {
